+++ milkman-jdbc/src/main/java/milkman/plugin/jdbc/domain/JdbcRequestContainer.java
@@ -1,10 +1,6 @@
 package milkman.plugin.jdbc.domain;
 
-import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 import milkman.domain.RequestContainer;
 
 @Getter @Setter
@@ -23,6 +19,10 @@ public class JdbcRequestContainer extends RequestContainer {
 	}
 
 
+	@Override
+	public RequestTypeDescriptor getTypeDescriptor() {
+		return new RequestTypeDescriptor("SQL", getClass().getResource("/icons/sql.png"));
+	}
 
 	public JdbcRequestContainer(String name, String jdbcUrl) {
 		super(name);