+++ milkman/src/main/java/milkman/domain/ResponseContainer.java
@@ -1,20 +1,16 @@
 package milkman.domain;
 
-import java.util.HashMap;
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonTypeInfo;
+import com.fasterxml.jackson.annotation.JsonTypeInfo.As;
+import com.fasterxml.jackson.annotation.JsonTypeInfo.Id;
+import lombok.*;
+
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
-import com.fasterxml.jackson.annotation.JsonTypeInfo.As;
-import com.fasterxml.jackson.annotation.JsonTypeInfo.Id;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
 
 @Data
 @AllArgsConstructor
@@ -26,7 +22,7 @@ public abstract class ResponseContainer {
 
 	
 	@JsonIgnore
-	private CompletableFuture<Map<String, String>> statusInformations = new CompletableFuture<>();
+	private CompletableFuture<Map<String, StyledText>> statusInformations = new CompletableFuture<>();
 	
 	
 	public <T extends ResponseAspect> Optional<T> getAspect(Class<T> aspectType) {
@@ -34,5 +30,24 @@ public abstract class ResponseContainer {
 				.filter(aspectType::isInstance)
 				.findAny()
 				.map(a -> (T)a);
+	}
+
+	@Getter @EqualsAndHashCode
+	public static class StyledText {
+		private final String text;
+		private final Optional<String> style;
+
+		public StyledText(String text) {
+			this(text, Optional.empty());
+		}
+
+		public StyledText(String text, String style) {
+			this(text, Optional.of(style));
+		}
+
+		private StyledText(String text, Optional<String> style) {
+			this.text = text;
+			this.style = style;
+		}
 	}
 }