+++ milkman/src/main/java/milkman/ui/main/ResponseComponent.java
@@ -12,11 +12,12 @@ import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.StackPane;
 import milkman.domain.RequestContainer;
+import milkman.domain.ResponseContainer.StyledText;
 import milkman.ui.components.FancySpinner;
 import milkman.ui.components.TinySpinner;
+import milkman.ui.main.options.CoreApplicationOptionsProvider;
 import milkman.ui.plugin.UiPluginManager;
 import milkman.utils.AsyncResponseControl;
-import milkman.utils.fxml.FxmlBuilder.*;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
@@ -86,12 +87,15 @@ public class ResponseComponent {
 	}
 
 
-	private void addStatusInformation(CompletableFuture<Map<String, String>> statusInformations) {
+	private void addStatusInformation(CompletableFuture<Map<String, StyledText>> statusInformations) {
 		statusDisplay.getChildren().clear();
 		statusInformations.thenAccept(stats ->  Platform.runLater(() -> {
-			for (Entry<String, String> entry : stats.entrySet()) {
+			for (Entry<String, StyledText> entry : stats.entrySet()) {
 				Label name = new Label(entry.getKey() + ":");
-				Label value = new Label(entry.getValue());
+				Label value = new Label(entry.getValue().getText());
+				if (!CoreApplicationOptionsProvider.options().isDisableColorfulUi()) {
+					entry.getValue().getStyle().ifPresent(value::setStyle);
+				}
 				value.getStyleClass().add("emphasized");
 				statusDisplay.getChildren().add(new HBox(name, value));	
 			}