+++ changelog.md
@@ -6,7 +6,8 @@ version 5.4.0 - upcoming
   * added option to not export keys on workspace export
   * added option to repeat a test n times
   * fix: ignore disabled headers in http exporter
-  * adde Drag&Drop for folders
+  * added Drag&Drop for folders
+  * added colors/icons to request types and status information (can be disabled in options)
 
 version 5.3.0
   * mask plain-keys on overview